backend/requirements-clean.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn==0.32.1
python-multipart==0.0.20

# Database and storage
supabase==2.11.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# AI and ML libraries
openai==1.58.1
pinecone-client==5.3.1

# Document processing
pypdf2==3.0.1
python-docx==1.1.3
beautifulsoup4==4.13.4
requests==2.32.3

# Data processing
pandas==2.2.3
numpy==2.1.3
pydantic==2.10.4

# Data ingestion
praw==7.8.1
feedparser==6.0.11

# Utilities
python-dateutil==2.9.0.post0
urllib3==2.3.0
